Atualizando a jlibcpp

+++ jio/include/jbitinputstream.h
@@ -66,19 +66,19 @@ class BitInputStream : public jio::InputStream{
 		 * \brief
 		 * 
 		 */
-		virtual long long Available();
+		virtual int64_t Available();
 
 		/**
 		 * \brief
 		 *
 		 */
-		virtual long long GetSize();
+		virtual int64_t GetSize();
 
 		/**
 		 * \brief
 		 *
 		 */
-		virtual long long GetPosition();
+		virtual int64_t GetPosition();
 		
 		/**
 		 * \brief
@@ -90,13 +90,13 @@ class BitInputStream : public jio::InputStream{
 		 * \brief
 		 * 
 		 */
-		virtual long long Read(char *, long long size);
+		virtual int64_t Read(char *, int64_t size);
     
 		/**
 		 * \brief Salto relativo.
 		 *
 		 */
-		virtual void Skip(long long skip);
+		virtual void Skip(int64_t skip);
 
 		/**
 		 * \brief
@@ -126,7 +126,7 @@ class BitInputStream : public jio::InputStream{
 		 * \brief
 		 *
 		 */
-		virtual long long GetReceiveBytes();
+		virtual int64_t GetReceiveBytes();
 
 };
