+++ jio/include/jdatainputstream.h
@@ -58,19 +58,19 @@ class DataInputStream : public jcommon::Object{
 		 * \brief
 		 * 
 		 */
-		virtual long long Available();
+		virtual int64_t Available();
 
 		/**
 		 * \brief
 		 *
 		 */
-		virtual long long GetSize();
+		virtual int64_t GetSize();
 
 		/**
 		 * \brief
 		 *
 		 */
-		virtual long long GetPosition();
+		virtual int64_t GetPosition();
 		
 		/**
 		 * \brief
@@ -100,7 +100,7 @@ class DataInputStream : public jcommon::Object{
 		 * \brief Salto relativo.
 		 *
 		 */
-		virtual void Skip(long long skip);
+		virtual void Skip(int64_t skip);
 
 		/**
 		 * \brief
@@ -118,7 +118,7 @@ class DataInputStream : public jcommon::Object{
 		 * \brief
 		 *
 		 */
-		virtual long long GetReceiveBytes();
+		virtual int64_t GetReceiveBytes();
 
 };
