+++ jio/include/jobjectoutputstream.h
@@ -57,7 +57,7 @@ class ObjectOutputStream : public jcommon::Object{
 		 * \brief
 		 *
 		 */
-		virtual long long Available();
+		virtual int64_t Available();
 
 		/**
 		 * \brief
@@ -69,7 +69,7 @@ class ObjectOutputStream : public jcommon::Object{
 		 * \brief
 		 * 
 		 */
-		virtual long long Flush();
+		virtual int64_t Flush();
 
 		/**
 		 * \brief
@@ -81,7 +81,7 @@ class ObjectOutputStream : public jcommon::Object{
 		 * \brief
 		 *
 		 */
-		long long GetSentBytes();
+		int64_t GetSentBytes();
 
 };
